requirements.txt
# FIXED requirements.txt - Compatible versions for beginners
# This version avoids dependency conflicts

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
sqlalchemy==1.4.53
psycopg2-binary==2.9.9

# Machine Learning (Simplified for beginners)
scikit-learn==1.3.2
transformers==4.36.0
torch==2.1.2

# Streaming & ETL (Simplified)
kafka-python==2.0.2
# apache-airflow==2.7.3  # Commented out - will add in Day 3 with proper setup
celery==5.3.4

# Database
redis==5.0.1
pymongo==4.6.0

# API & Web
httpx==0.25.2
python-multipart==0.0.6

# Monitoring (Basic)
prometheus-client==0.19.0

# Development Tools
pytest==7.4.3
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.8.2
requests==2.32.3

yfinance==0.2.28
pandas-datareader==0.10.0

# Add these to your existing requirements.txt
torch==2.1.2
transformers==4.36.0
scikit-learn==1.3.2
tensorflow==2.15.0
mlflow==2.8.1
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
joblib==1.3.2
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2